Initial commit of Hiera ENC
This code is based off the original work from @Ghetzel. In light of
Puppet 3's Hiera integration and a few lessons learned about the
security of Facter facts, this code rewrite is being done. Another goal
already partially schived by this module is to seperate this ENC from
the now dying hiera-puppet module. With the inclusion of Hiera in Puppet
core, that module is no longer a good place to keep this code.