First pass refactor
This is the first pass of refactoring.

Big changes include the removal of Hiera setup method. The lookup
functions have been changed to use the HieraPuppet facilities now
present in Puppet 3. Creation of Hiera::Scope has been removed as well
as it is handeled in the new HieraPuppet. Some function names have been
changed in anticipation of the rest of the refactor.